[refactor] storybook 피드백 사항 해결
1. stories 폴더에는 story 관련 파일만
2. Icons.stories.tsx,  Inputs.stories.tsx 파일  코드 리팩토링
