+++ src/styles/index.ts
@@ -2,3 +2,4 @@ export { default as colors } from "./colors";
 export { default as fonts } from "./fonts";
 export { default as global } from "./global";
 export { default as mixin } from "./mixin";
+export { default as sizes } from "./sizes";
